Moves BetaWarning.js into src/components
