+++ src/MainMenu.js
@@ -6,7 +6,7 @@ import {
   Menu,
 } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
-import { BetaWarning } from './BetaWarning';
+import { BetaWarning } from './components/BetaWarning';
 import logo from './images/logo.svg';
